@@ -4,9 +4,11 @@
   "private": true,
   "dependencies": {
     "cra-template": "1.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.1"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.0.2",
+    "react-scripts": "5.0.1",
+    "sass": "^1.83.0"
   },
   "scripts": {
     "start": "react-scripts start",
