@@ -1,9 +1,18 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
+import ReactDOM from 'react-dom/client';
+import './index.scss';
+import { BrowserRouter } from 'react-router-dom';
+import AllRoutes from './routes.jsx';
+import Header from './components/multiPageComponents/header.jsx';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
-  </React.StrictMode>,
-  document.getElementById('root')
+    <BrowserRouter>
+
+    <Header/>
+    <AllRoutes />
+    
+    </BrowserRouter>
+  </React.StrictMode>
 );
